refactor: added ResetPasswordPresenter and ResetPasswordView to comply with MVP (#160)
